+++ lib/ui/widgets/auth/auth_widget.dart
@@ -42,26 +42,4 @@ class AuthWidget extends StatelessWidget {
 
 
 
-class _ErrorMessageWidget extends StatelessWidget {
-  const _ErrorMessageWidget({Key? key}) : super(key: key);
 
-  @override
-  Widget build(BuildContext context) {
-    final errorMessage = context.select((AuthViewCubit c) {
-      final state = c.state;
-      return state is AuthViewCubitErrorState ? state.errorMessage : null;
-    });
-    if (errorMessage == null) return const SizedBox.shrink();
-
-    return Padding(
-      padding: const EdgeInsets.only(bottom: 20),
-      child: Text(
-        errorMessage,
-        style: const TextStyle(
-          fontSize: 17,
-          color: Colors.red,
-        ),
-      ),
-    );
-  }
-}
