+++ lib/ui/widgets/auth/components/form_widget.dart
@@ -1,5 +1,6 @@
 import 'package:comics_db_app/domain/services/auth_data_storage.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 
 class FormWidget extends StatelessWidget {
   FormWidget({Key? key}) : super(key: key);
